Checkpoint Event

+++ db/sqlc/dashboard-profile.sql.go
@@ -28,11 +28,11 @@ type HomeDashboardRow struct {
 	NamaLengkap       sql.NullString `json:"nama_lengkap"`
 	UrlFoto           sql.NullString `json:"url_foto"`
 	University        sql.NullString `json:"university"`
-	FinishedPractice  sql.NullInt32  `json:"finished_practice"`
+	FinishedPractice  sql.NullInt64  `json:"finished_practice"`
 	JumlahPractice    int64  `json:"jumlah_practice"`
-	FinishedChallange sql.NullInt32  `json:"finished_challange"`
+	FinishedChallange sql.NullInt64  `json:"finished_challange"`
 	JumlahChallange   int64  `json:"jumlah_challange"`
-	TotalSkor         sql.NullInt32  `json:"total_skor"`
+	TotalSkor         sql.NullInt64  `json:"total_skor"`
 }
 
 func (q *Queries) HomeDashboard(ctx context.Context, idMahasiswa int32) (HomeDashboardRow, error) {
